Add ShortShape to all the key places to get @range support

+++ codegen-server/src/test/kotlin/software/amazon/smithy/rust/codegen/server/smithy/ConstrainedShapeSymbolProviderTest.kt
@@ -49,7 +49,7 @@ const val baseModelString =
     string ConstrainedString
 
     @range(min: -2, max: 10)
-    integer ConstrainedShort
+    short ConstrainedShort
 
     @range(min: 10, max: 29)
     integer ConstrainedInteger
